[svn-r15237] Description: 	Merge revisions 14900:15037 from trunk into metadata journaling branch
Tested on:
        FreeBSD/32 6.2 (duty) in debug mode
        FreeBSD/64 6.2 (liberty) w/C++ & FORTRAN, in debug mode
        Linux/32 2.6 (kagiso) w/PGI compilers, w/C++ & FORTRAN, w/threadsafe,
                                in debug mode
        Linux/64-amd64 2.6 (smirom) w/default API=1.6.x, w/C++ & FORTRAN,
                                in production mode
        Linux/64-ia64 2.6 (cobalt) w/Intel compilers, w/C++ & FORTRAN,
                                in production mode
        Solaris/32 2.10 (linew) w/deprecated symbols disabled, w/C++ & FORTRAN,
                                w/szip filter, in production mode
        Mac OS X/32 10.5.2 (amazon) in debug mode
        Linux/64-ia64 2.4 (tg-login3) w/parallel, w/FORTRAN, in production mode

+++ src/H5Oprivate.h
@@ -337,8 +337,8 @@ typedef struct H5O_layout_contig_t {
 typedef struct H5O_layout_chunk_t {
     haddr_t	addr;			/* File address of B-tree            */
     unsigned	ndims;			/* Num dimensions in chunk           */
-    size_t	dim[H5O_LAYOUT_NDIMS];	/* Size of chunk in elements         */
-    size_t      size;                   /* Size of chunk in bytes            */
+    uint32_t	dim[H5O_LAYOUT_NDIMS];	/* Size of chunk in elements         */
+    uint32_t    size;                   /* Size of chunk in bytes            */
     H5RC_t     *btree_shared;           /* Ref-counted info for B-tree nodes */
 } H5O_layout_chunk_t;
